Move shell-arg replacing into utils
